+++ packages/riverpod_analyzer_utils/lib/src/riverpod_ast/provider_listenable_expression.dart
@@ -11,7 +11,7 @@ class ProviderListenableExpression extends RiverpodAst {
   });
 
   @internal
-  static ProviderListenableExpression? parse(Expression? expression) {
+  static ProviderListenableExpression? _parse(Expression? expression) {
     if (expression == null) return null;
 
     // print('oy $expression // ${expression.runtimeType}');